pyproject.toml
[tool.poetry]
name = "scraper-bot"
version = "0.1.0"
description = ""
authors = ["Koosha Sangari <ko.sangari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "4.11.1"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
idna = "3.4"
pydantic = "1.10.7"
pytelegrambotapi = "4.8.0"
python-dotenv = "0.21.0"
requests = "2.28.1"
schedule = "1.1.0"
soupsieve = "2.3.2.post1"
urllib3 = "1.26.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
